estimate memory footprint for some structures
We would like to estimate, as closely as possible, how much memory does
a memtable use in the LSA.

The first thing we need to look at, is how much memory is being held by
managed_bytes, managed_vectors, etc. However, code like this:

    memtable_entry* entry = current_allocator().construct<memtable_entry>(
        _schema, dht::decorated_key(key), mutation_partition(_schema));

allocates memory in the LSA as well. That happens for some key structures in the
life of a memtable. If the partitions and the data are big, we could very well
just ignore this overhead. But if they aren't, then ignoring it may lead to
large errors.

In particular, if the data is so small that it is less than the size external
size threshold for managed_bytes, we'll account 0 bytes, but the memtable entry
will use more than 200 bytes. In a memtable that is full of data like that,
we'll scan the whole of it accounting for 0 bytes - when in reality it can be
using a lot of memory.

Since we are just dealing with the size of structures here, technically a simple
sizeof() will do. However, since what exactly is allocated through LSA is a
decision that is internal to the class at hand, I'd prefer to encapsulate it.

Signed-off-by: Glauber Costa <glauber@scylladb.com>