sstables: keep track of a lower bound on memory written
This patch keeps track of how much memory out of a memtable we have already
written to the disk buffers. It will be used later for virtual dirty memory
accounting to estimate how close we are to finish writing a memtable to its
fullest.

For virtual dirty to yield the best possible results, we must be as close as
possible to the real memory usage. But for simplicity, I am leaving the LSA
object overhead out of it.

That is (at this time) 14 bytes per object plus padding, which is the size of
the object_descriptor object in logalloc.cc.

It is technically possible to compute this information, but that would require
us to hold a reference to the region somewhere in the SSTable writer. The
complexity added by this simply does not justify that.

Still, the name of the method will reflect that this is a lower bound.

Signed-off-by: Glauber Costa <glauber@scylladb.com>