sstables: allow a callback to be called at every partition written
For the virtual dirty memory management to work, we need to communicate
somehow to the memtable writer layer that we are somehow progressing.

One way of doing this is to pass a pointer to the necessary objects (potentially
memtable, at least the region group) to the SSTable object. Another is to just
allow us to register a callback that the SSTable object will call at the right
time.

This patch takes route #2. My main justification for doing this, is that doing
otherwise would have forced us to potentially create entanglements between the
SSTable and other objects that the SSTable shouldn't even necessarily know
about. It would also bring all the virtual dirty logic inside the SSTable.

Signed-off-by: Glauber Costa <glauber@scylladb.com>