partition_snapshot_reader: do not call do_fill buffer in the constructor
That causes virtual functions to be incorrectly assigned. If they are called in
the constructor they will be bound to the version that exists at the moment - in
this class itself - without derived functions being given the chance to
specialize it.

Turns out we don't really have any reason to call do_fill_buffer() in the
constructor. Upon seeing the buffer empty at next use, we will fill it.

Signed-off-by: Glauber Costa <glauber@scylladb.com>