allow looking at the range_tombstone_stream without converting to a mutation_fragment
Currently, code interacting with range_tombstone_streams uses the get_next() API
to convert the next available tombstone in the list to a mutation fragment.

In order to properly account the range tombstone, it is desirable to do it
earlier, since the creation of the mutation fragment may move the data to a
different region.

This patch introduces the peek_next() API, which allows one to obtain a pointer
to the next element in the list, and implements the get_next() API in terms of
it.

Signed-off-by: Glauber Costa <glauber@scylladb.com>