commitlog: remove must_sync() condition
must_sync() checks whether or not we should sync, using as a criteria the time
elapsed since last sync. That can never be more than twice the maximum defined
time in the config file.

However, we already register a timer that fires periodically at the maximum
defined sync time. Therefore, under normal conditions this code path is never
executed.

If the system piles up with load, it might be that the timer gets delayed. But
if that's the case, forcing a sync ourselves will only add to the load.

This code seems to predate the timer mechanism according to git history, which
is the only justification for its existence. But in the current form, it only
seems to be adding complexity.

Signed-off-by: Glauber Costa <glauber@scylladb.com>