commitlog: rewrite logic for waiting on pending allocations
The current logic for when and how to wait on a request leaves a lot of
room for improvement. That's not only a theoretical problem, as we have
seen reports of very frequent timeouts that can render relatively simple
workloads useless.

There are a number of problems with the current code: first, the latencies
introduced are what we need to flush the current buffer. But there is no need
to wait for the current buffer to finish. We should be able to make progress
when any buffer finishes, which is what we do here.

Second, the criteria for blocking are quite arbitrary. For requests of a larger
size, maintaining the number of in flight requests constant will lead to a very
large amount of memory piling up. Even if we were to split the buffers, the fact
that we need the full cycle to end does not allow us to properly take advantage
of it. A memory-based approach, which is what we take here, guarantees that the
amount of pending memory.

Signed-off-by: Glauber Costa <glommer@scylladb.com>