commitlog: split large buffers into multiple small ones when writing
Although buffers are 128 kB by default, they can always be bigger than that. If
we have a very large mutation, that will force the buffer to increase and it
will be at least as large as the mutation.

When we cycle those buffers, we will be generating very large writes. This did
not play well with the old max_active_writes-based code at all, because we would
find ourselves suddenly holding onto a constant number of now very large
buffers, and those could take a lot of time to fully cycle.

But even after that code is replaced, writing the bigger buffers is still not a
good idea, for mainly two reasons:

1) We will need the whole buffer to complete before we can allow more requests.
   If a big request takes us over limit but the requests that follow are
   smaller, we can start servicing them earlier if we are finishing the writes
   piecemeal.

2) The Linux fs layer doesn't like that many big buffers in flight. Figures for
   system iowait go as high as 25 % (observed in my system), whereas with this
   patch applied they are kept low (not zero unfortunately, but that is a
   different story).

Signed-off-by: Glauber Costa <glommer@scylladb.com>