correctly calculate latencies for writes
Right now we are calculating latencies only when we are about to add an
item to the memtable.

That's incorrect and misleading, for two reasons. First, it leaves the
commitlog latencies out. But second, it is done after the memtable wall
effect is applied, which means we are not counting throttle time neither
in the memtables or in the commitlog.

To do that, we'll start the latency_counter object as soon as possible
and move it after apply_in_memory(). That should span the entire write
operation.

We will change the signature of apply_in_memory() so we can move it as
late as possible. There's a precedent in that: that's how the storage
proxy calculates its latencies. In previous versions I have moved this
to a continuation attached to apply_in_memory(). However, we no longer
have such continuation, so it's expensive to add one just for that.

Signed-off-by: Glauber Costa <glauber@scylladb.com>
---
v2: rebased for master
v3: "it's just a quick change, no need to recompile"
v4: refactor do_apply so we don't need to change apply_in_memory
v5: special interface to return a pointer to cf; don't add an extra
    continuation for the latency measurement in fast path.
v6: rebased after many changes in master. There is no longer a
    continuation in apply_in_memory() so change its signature instead.
    No need to add an interface to fetch the CF pointer.