correctly calculate latencies for writes
Right now we are calculating latencies only when we are about to add an
item to the memtable.

That's incorrect and misleading, for two reasons. First, it leaves the
commitlog latencies out. But second, it is done after the memtable wall
effect is applied, which means we are not counting throttle time neither
in the memtables or in the commitlog.

To do that, we'll start the latency_counter object as soon as possible
and move it after apply_in_memory(). That should span the entire write
operation.

Signed-off-by: Glauber Costa <glauber@scylladb.com>
---
v2: rebased for master
v3: "it's just a quick change, no need to recompile"
v4: refactor do_apply so we don't need to change apply_in_memory