commitlog: serialize incoming mutations
We need to serialize incoming mutations in the commitlog context, to
make sure that they will have increasing replay positions.

The database code will test for such reorders, and in case a reorder is
seen, issue the request again. There is mention that those reorders are
supposed to be rare, but they aren't: at least when the request sizes
are big enough.

If we do reinsert the requests, timeouts are almost bound to happen. Big
requests already take a lot of time to complete just due to their size.
If we have to wait for them to complete multiple times, it's almost
certain we'll have to go over the timeout cap.

With this patch, a workload of 64kB-sized writes that would timeout with
100 threads happily goes over 500 threads without hiccups.

Signed-off-by: Glauber Costa <glauber@scylladb.com>