database: add a get_column_family signature, returning a lw_shared_ptr
There are places in which we need to use the column family object many
times, with deferring points in between. Because the column family may
have been destroyed in the deferring point, we need to go and find it
again. That makes the use of the existing find_column_family a bit
inconvenient in those places.

If we use lw_shared_ptr, however, we'll be able to at least guarantee
that the object will be alive. For that to happen, we will introduce a
new method, get_column_family(), that returns the lw_shared_ptr.

It is named consistently to get_column_families(), which will return
pointers to all column families in the database.

Signed-off-by: Glauber Costa <glauber@scylladb.com>