sstables: allow us to wait on deleting sstables to complete.
When we destroy an sstable, we trigger some asynchronous work without
waiting for it. That work is needed for the existing shards to
coordinate the process of deleting this table.

There are operations, however, that shouldn't be considered done until
that work is quiesced. We need, then, a way to wait on those futures.
This patch introduces such a way. It allow us to wait on all sstables
that were deleted up until this moment.

Signed-off-by: Glauber Costa <glauber@scylladb.com>