database: start flushing a memtable before the previous one finishes
We need to start a flush before the current one finishes, otherwise
we'll have a period without significant disk activity when the current
SSTable is being sealed, the caches are being updated, etc.

By design, every time a memtable is finished flushing it will call
revert_potentially_cleaned_up_memory() to revert the virtual memory
charges.  That is the perfect moment for us to act. It indicates that
all the data flushing part is done.

The way we'll do it is by keeping the semaphore_units alive for this
memtable. When the flush ends, we destroy that object. This will
effectively trigger the next flush if there is a next flush that can be
initiated.

Signed-off-by: Glauber Costa <glauber@scylladb.com>