database: avoid infinite recursion with replay position reordering
We have work pending to avoid replay position reorderings altogether.
While that doesn't happen, we are still in a fragile position in that
reorder events can cause us to loop forever into an infinite recursion
if they keep happening.

This patch temporarily fixes that by establishing a maximum number of
attempts. After that we'll give up and throw. Hopefuly we'll be able to
get rid of it by making reorders either impossible or irrelevant.

Signed-off-by: Glauber Costa <glauber@scylladb.com>