database: flush oldest memtable instead of largest
We are currently flushing the biggest memtable in a region group, but
that might not be the best pick.

On top of that, the current criteria is not being respected.  The
problem is demonstrated to happen when the dirty memory group is over
limit, and so is the system table extra allowance. In that situation,
both the normal region and the system table region will be under
pressure and try to flush.

Due to the hierarchical nature of the region_group, the system region
will pick a memtable to flush from the normal region, since it has a
best match for the flush criteria. However, since the normal region is
also under pressure, it will also pick a memtable itself. As the very
large memtable is out of the way, flushed by the system region, the
normal region will now pick a very small one and proceed.

This problem is by no means exclusive to system tables, and can happen
with streaming as well. The fix proposed ended up being a bit
convoluted, and while discussing it we have also managed to spot a few
other issues with the hierarchical approach, A hightlight among them
being a priority inversion problem where streaming memtables and normal
memtables now affect each other's priority.

In the face of that, it would be better to just go directly where we
intend to, and move the criteria to flush the memtable with the oldest
mutation.

Tracking that is easy: when a memtable gets its first mutation it's put
in a (ideally) FIFO queue, and when it's time to flush we just pick the
front of the queue. The actual solution uses an intrusive list, because
explicit flushes can trigger the flush of specific column families
regardless of their position in the list. So we use this as a FIFO-like
queue that support removal of the middle elements.

There is nice performance improvement from this patch, but that's mostly
due to the fact that the previous criteria wasn't being respected.

Fixes #1854

Signed-off-by: Glauber Costa <glauber@scylladb.com>