database: rework dirty memory hierarchy
Even after we change the criteria of memtable flush to flush-the-oldest,
there are still problems that are not solved. More specifically, because
the normal region inherits from the system region, if the normal region
is under pressure (over the soft limit threshold), the system region
will certainly be as well, even though it has an extra allowance. This
is because after virtual dirty, we start blocking when we reach half the
region, but memory itself can grow up to 100 % of the region. So the
total amount of memory used will be certainly bigger than the system
pressure threshold, which is now 50 % plus the allowance.

To fix that, this patch reorganizes the dirty memory region groups as
follows:

The total amount of dirty memory is divided up into three groups:
system, normal, and streaming. They are organized in a tree as siblings
sitting below the root, the _overall_dirty_memory_container. That
container has a limit that is equal to _memtable_total_space plus a
system allowance, and it will guarantee that when the sum of the
children's memory reach that limit, we'll throttle.

The overall_dirty_memory_container doesn't really hold any memtables and
just work as a container. For that reason, there is no need to apply any
virtual dirty considerations to it.  It also doesn't ever flush. For
that to work without freezing the system we need to guarantee that at
least one group below it will have triggered a flush already when the
overall_dirty_memory_container reaches its limit.

Since all regions have a soft limit threshold that triggers the flush,
we can calculate what should the maximum threshold be to guarantee that
condition. Flushes have to happen at the latest when the region fills up
to 44 % of their size (recall that streaming is 4 times smaller, 0.44 +
0.44 + 0.11 = 99 %)

The other two regions define both a limit and a soft limit at which the
flushing must start

Fixes #1918

Signed-off-by: Glauber Costa <glauber@scylladb.com>