LSA: add age-based tracking
Allow regions to be tracked based on which of them holds the oldest segment.
This can be used to implement flush-the-oldest for memtables as described in

I had a previous attempt where the oldest memtable is tracked within the
region itself which is fairly simple and easy. However, with such
implementation we would have to always compare all children's timestamps
from a parent group if we are to respect the criteria when flushing from
a parent region.

This patch uses the existing LSA infrastructure for tracking of maximal
regions, but allows tracking to happen on an age-based manner, where the
oldest region is the maximal region.

To keep the code simple we want the oldest region to have a higher
number, so instead of using timestamps we just decrement a counter in
vector-clock style. That's also much cheaper than getting the timestamp
anyway.

Signed-off-by: Glauber Costa <glauber@scylladb.com>