database: rework dirty memory hierarchy
Issue #1918 describes a problem, in which we are generating smaller
memtables than we could, and therefore not respecting the flush
criteria.

That happens because the normal region inherits from the system region,
if the normal region is under pressure (over the soft limit threshold),
the system region will certainly be as well, even though it has an extra
allowance. This is because after virtual dirty, we start blocking when
we reach half the region, but memory itself can grow up to 100 % of the
region. So the total amount of memory used will be certainly bigger than
the system pressure threshold, which is now 50 % plus the allowance.

This is a direct consequence of the linear hierarchy between the regions
and to guarantee that it won't happen we would have to play games with
regions picking one another's semaphore. While that works, it can lead
to problems on its own, like priority inversion if the regions have
different priorities - like streaming and regular.

To fix that, this patch reorganizes the dirty memory region groups as
siblings of each other under a container parent. Because the container
parent has no memtables on its own, the regions can easily just
consume() units in the parent semaphore. That is simpler to handle and
do not prevent regions that should flush in parallel with their own
priority to do so.

Some explanation about this is due: the linear hierarchy existed for a
reason and it had properties that we would like to see respected. In
particular we want any group to be able to use its full allowance when
the other groups are not active, but share the total memory when more
than one is active.

That can also be achieved with a siblings-with-container hierarchy: each
region defines its limits in the same way as before, and the container
defines its limit to be the maximum of all three limits.

Fixes #1918

Signed-off-by: Glauber Costa <glauber@scylladb.com>